Fixed the Maven Formatting
Took a look at official Maven repos (like Gson's, all of Google's, and the NOVA Maven repo) and saw that this was the format that was used.

Also: Do you expect people to put try and catch blocks in an annotation? Hell, no. (Changed the URLs in MavenDependency to Strings)