Write more tests + change return type of ItemIngredient.consumeOnCrafting to Optional.
The reason for the return type change is that before the return was
a type of Item, which had to be non-null. This would require additional
checks in CraftingRecipe to see if the result had an item count of 0
and remove the item from the grid if that was the case.

Another reason is that the CraftingGrid.setStack takes an Optional
as the item parameter.